device/requirements.txt
azure-iot-device
gpiozero
jsonpickle